requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
qrcode[pil]==7.4.2
Pillow==10.0.1